chore(docusaurus): swap scarf for reo
