+++ package.json
@@ -22,7 +22,7 @@
     "@shikijs/rehype": "~3.4.2",
     "@shikijs/transformers": "~3.4.2",
     "@wasmcloud/docusaurus-github-stars": "0.0.0",
-    "@wasmcloud/docusaurus-scarf-analytics": "0.0.0",
+    "@wasmcloud/docusaurus-reo-analytics": "0.0.0",
     "@wasmcloud/docusaurus-seo-checks": "0.0.0",
     "clsx": "^2.1.1",
     "react": "^19.1.0",
