+++ packages/docusaurus-reo-analytics/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@wasmcloud/docusaurus-scarf-analytics",
+  "name": "@wasmcloud/docusaurus-reo-analytics",
   "version": "0.0.0",
   "private": true,
   "description": "",
